common/results.py: rework handling of signals
Throwing an exception out of signal handler was not the correct way to
properly terminate the command executed in foreground.  After the rework
the signal handler forwards the signal to the subprocess executed by
exec_cmd() or get_cmd_output() and sets the exit code appropriately.
The exception is then thrown by the foreground process in those methods.